#38 move AppScreens to new file - fix some code style. - add room schemaLocation for Ksp - update some dependency
