+++ app/src/main/java/ir/thatsmejavad/backgroundable/NavGraph.kt
@@ -5,6 +5,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.NavType
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
+import ir.thatsmejavad.backgroundable.core.AppScreens
 import ir.thatsmejavad.backgroundable.core.viewmodel.daggerViewModel
 import ir.thatsmejavad.backgroundable.screens.collectionlist.CollectionListScreen
 import ir.thatsmejavad.backgroundable.screens.mediadetail.MediaDetailScreen
@@ -88,21 +89,5 @@ internal fun NavGraphBuilder.mainNavGraph(navController: NavHostController) {
             viewModel = daggerViewModel(),
             onBackClicked = { navController.navigateUp() }
         )
-    }
-}
-
-internal sealed class AppScreens(val route: String) {
-    object CollectionList : AppScreens("collection-list")
-
-    object MediaList : AppScreens("media-list?id={id}&title={title}") {
-        fun createRoute(id: String, title: String): String {
-            return "media-list?id=$id&title=$title"
-        }
-    }
-
-    object MediaDetail : AppScreens("media-detail?id={id}&title={title}") {
-        fun createRoute(id: Int, title: String): String {
-            return "media-detail?id=$id&title=$title"
-        }
     }
 }
